fix: move function `spanTreeEntries` to its own file
