+++ clients/web/src/lib/span/get-span-children.ts
@@ -21,33 +21,3 @@ export function getSpanChildren(
 
   return all;
 }
-
-export type TreeEntry = {
-  depth: number;
-  span: Span;
-};
-
-export function treeEntries(
-  span: Span,
-  filter: (span: Span) => unknown = () => true,
-  maxDepth = 10,
-  depth = 0,
-): TreeEntry[] {
-  const self = { depth, span };
-  const showMe = filter(span);
-  if (depth === maxDepth) return showMe ? [self] : [];
-
-  // Ensure children are chronologically ordered before recursion.
-  const children = [...span.children].sort((a, b) => a.createdAt - b.createdAt);
-
-  // Concat each of the child trees.
-  return children.reduce(
-    (agg, v) =>
-      agg.concat(
-        // Note: we don't increase depth if we're filtering the current node.
-        // This way we visually "reassign" our parent to one that matches the level filter.
-        treeEntries(v, filter, maxDepth, showMe ? depth + 1 : depth),
-      ),
-    showMe ? [self] : [],
-  );
-}
